V 1.5.15
Millores:
Al reimprimir ja surt la zona correcta, aquest canvi només és a futur, és a dir, les denúncies antigues continuaran dependent de la seleccionada, però les fetes amb aquesta nova versió ja mantindran les seves pròpies zones.

Canvis

El text de les fotos ara conte el numero de denuncia, i s'ha reduït el tamany de la lletra.
Solventat un error on el numero de carrer al reimprimir una denuncia sortia amb un ".0" adicional.

+++ app/src/main/java/com/boka2/gesblue/activities/RecuperarDenunciaActivity.java
@@ -178,7 +178,7 @@ public class RecuperarDenunciaActivity extends AppCompatActivity implements Cust
             mSancio = new Sancio(denuncia.getMatricula(),"",tipusVehicle,marca,model,color,infraccio,carrer,zona);
         }
         else{
-            mSancio = new Sancio(denuncia.getMatricula(),String.valueOf(denuncia.getAdrecanum()),tipusVehicle,marca,model,color,infraccio,carrer,zona);
+            mSancio = new Sancio(denuncia.getMatricula(),String.valueOf(((int)denuncia.getAdrecanum())),tipusVehicle,marca,model,color,infraccio,carrer,zona);
         }
         intent.putExtra(ReimpressioTiquet.INTENT_SANCIO, mSancio);
         intent.putExtra(ReimpressioTiquet.INTENT_NUM_DENUNCIA,denuncia.getCodidenuncia());
